Android: Now using 16-bit integers instead of 32-bit integers to carry wave data from PlayerService to EqualizerPresetsPresenter. Added shadow for text and GPU optimization in OutputMeterView.
Related to issue #406.

+++ MPfm/MPfm.MVP/Presenters/EqualizerPresetsPresenter.cs
@@ -128,7 +128,7 @@ namespace MPfm.MVP.Presenters
                 }
                 else
                 {
-                    Tuple<int[], int[]> data = _playerService.GetMixerData(0.02);
+                    Tuple<short[], short[]> data = _playerService.GetMixerData(0.02);
 
                     // Convert to floats (TODO: Try to optimize this. I'm sure there's a clever way to do this faster.
                     float[] left = new float[data.Item1.Length];
